requirements.txt
epftoolbox @ git+https://github.com/jeslago/epftoolbox.git@47d6e0629f65ebd19d3c12cb5689dbad0c2ea078
finta==1.3
keras-tuner==1.4.7
matplotlib==3.10.0
numpy==2.0.2
optuna==4.5.0
optuna-integration==4.5.0
pandas==2.2.2
seaborn==0.13.2
sklearn-pandas==2.2.0
tensorflow==2.19.0
tqdm==4.67.1